src/config.py
import os
from dotenv import load_dotenv

# load environment variables from .env file
load_dotenv()

# access environment variables
AWS_ACCESS_KEY_ID = os.getenv('AWS_ACCESS_KEY_ID')
AWS_SECRET_ACCESS_KEY = os.getenv('AWS_SECRET_ACCESS_KEY')
