requirements.txt
librosa
moviepy
numpy
scipy
ffmpeg-python
flask
flask-cors
boto3
python-dotenv